+++ src/main.c
@@ -4,9 +4,9 @@
 #include <SDL2/SDL.h>
 #include "devm.h"
 #include "disp.h"
-#include "file.h"
 #include "harv.h"
 #include "hold.h"
+#include "home.h"
 #include "loot.h"
 #include "main.h"
 #include "make.h"